refactor: fix ESLint errors
