+++ lib/rewardops.js
@@ -5,7 +5,7 @@ const emitter = require('./emitter');
 const { program, programs } = require('./resources/programs');
 const { setLogFilePath } = require('./utils/logger');
 const urls = require('./urls');
-const { version } = require('../package');
+const { version } = require('../package.json');
 
 /**
  * RewardOps Node SDK global object.
