+++ lib/utils/logger.js
@@ -14,7 +14,7 @@ const redact = require('redact-secrets');
 const Redactyl = require('redactyl.js');
 const { serializeError } = require('serialize-error');
 
-const config = require('./../config');
+const config = require('../config');
 const { LOG_PREFIX } = require('../constants');
 
 const { combine, timestamp: timestampFormatter, printf } = winston.format;
